Creating Update and Delete request, only post author can delete or update post

+++ src/module/posts/posts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, Post } from '@nestjs/common';
 import { ExceptionCodeName } from 'src/enum/exception-codes.enum';
 import { getRepository } from 'typeorm';
 import { RequestUserPayload } from '../auth/interface/request-user-payload.interface';
@@ -16,21 +16,7 @@ export class PostsService {
     private userService: UserService
   ){}
 
-  /*async create(
-    createPostDto: CreatePostDto,
-  ): Promise<PostModel> {
-    
-    const { title, content, date } = createPostDto;
-
-    const post = new PostModel();
-    post.title = title;
-    post.content = content;
-    post.date = date;
-    //return getRepository(Plant).save(plant);
-    return post;
-  }*/
-  //@Transactional()
-  
+  @Transactional()
   async create(
       createPostDto: CreatePostDto,
       requestUserPayload: RequestUserPayload,
@@ -54,46 +40,81 @@ export class PostsService {
     return getRepository(PostModel).save(post);
   }
   
-  /*
+  
   @Transactional()
-  async update(id: number, updatePlantDto: UpdatePlantDto): Promise<Plant> {
-    const { name, info, image_path, days_water, care } = updatePlantDto;
-    const plant = await this.getOneById(id);
-    plant.name = name;
-    plant.care = care;
-    plant.info = info;
-    plant.image_path = image_path;
-    plant.days_water = days_water;
-    return getRepository(Plant).save(plant);
+  async update(
+    id: number, 
+    requestUserPayload: RequestUserPayload,
+    updatePostDto: UpdatePostDto,
+  ): Promise<PostModel> {
+    const user = await this.userService.getOne({
+      where: {
+        id: requestUserPayload.id,
+      },
+    });
+    if (!user) {
+      throw new UnauthorizedException(ExceptionCodeName.INVALID_CREDENTIALS);
+    }
+    const user_id = user.id;
+    const post = await this.getOneById(id, user_id);
+    if (!post) {
+      throw new NotFoundException(ExceptionCodeName.YOU_CAN_NOT_UPDATE_THIS_POST);
+    }
+    
+    const { title, content, date } = updatePostDto;
+    const newPost = {...post};
+    newPost.title = (title) ? title : newPost.title;
+    newPost.content = (content) ? content : newPost.content;
+    newPost.date = (date) ? date : newPost.date;
+
+    return await getRepository(PostModel).save(newPost);
   }
-  */
   
   
   @Transactional()
   async getAll(): Promise<PostModel[]> {
-    return getRepository(PostModel).createQueryBuilder('post_model').getMany();
+    return getRepository(PostModel).createQueryBuilder('post_model')
+    .leftJoinAndSelect("post_model.user", "user")
+    .getMany();
   }
 
-  /*
+  
   @Transactional()
-  async getOneById(id: number): Promise<Plant> {
-    const res = await getRepository(Plant)
-      .createQueryBuilder('plant')
-      .where('plant.id = :id', { id })
-      .getOne();
-    if (!res) {
-      throw new NotFoundException(ExceptionCodeName.INVALID_PLANT_ID);
+  async delete(id: number, requestUserPayload: RequestUserPayload): Promise<void> {
+    const user = await this.userService.getOne({
+      where: {
+        id: requestUserPayload.id,
+      },
+    });
+    if (!user) {
+      throw new UnauthorizedException(ExceptionCodeName.INVALID_CREDENTIALS);
     }
-    return res;
-  }
-
-  @Transactional()
-  async delete(id: number): Promise<void> {
-    getRepository(Plant)
+    const user_id = user.id;
+    const post = await this.getOneById(id, user_id);
+    if (!post) {
+      throw new NotFoundException(ExceptionCodeName.YOU_CAN_NOT_DELETE_THIS_POST);
+    }
+        
+    getRepository(PostModel)
       .createQueryBuilder()
       .delete()
       .where('id = :id', { id })
+      .andWhere('user_id = :user_id', {user_id})
       .execute();
   }
-  */
+
+
+  @Transactional()
+  async getOneById(id: number, user_id: number): Promise<PostModel> {
+    const res = await getRepository(PostModel)
+      .createQueryBuilder('post_model')
+      .where('id = :id', { id })
+      .andWhere('user_id = :user_id', {user_id})
+      .getOne();
+    if (!res) {
+      throw new NotFoundException(ExceptionCodeName.POST_DOES_NOT_EXIST_OR_YOU_DO_NOT_OWN_THIS_POST);
+    }
+    return res;
+  }
+  
 }
